Remove useless dependency on build tasks
Previously, teavmc task declared a dependency to the build task of
the user project. This was useless because TeaVM only needs compiled
class files, it doesn't need jars or distribution archives.

This was very costly as these useless tasks represented about half
of the build time.

This commit removes this dependency on useless tasks, to double the
build speed.