Add incremental build capacity
Previously, using the teavm-gradle-plugin meant that TeaVM compilation
happened at every single build where the task was a dependency. Gradle
never considered the task up to date. This resulted in an enormous
amount of time lost uselessly. For example, running `./gradlew :backend:test`
would trigger compilation by TeaVM.

This commit fixes it by declaring inputs and outputs of the TeaVM task. As
such it is skipped if nothing has changed in places that concern TeaVM, and
only executes when needed.