Add TeaVM Metaprogramming as compile dependency of user project
Previously, user project applying this plugin would not get the
metaprogramming API of TeaVM included as a dependency. This made
it either impossible to use it, or required user to understand
what it is, that it is a separate module, and that they need to
add it themself. This was an obstacle in the way of a smooth
start with TeaVM. This was among other things required to use
Flavour.

This commit fixes it by adding the Metaprogramming module to
user project's dependencies.