Add support for Flavour
Previously, if user wanted to use Flavour, they would have to add the
corresponding dependencies themself. This meant more work for user,
less smoothness, more obstacles in the way of an easy start with TeaVM.

This commit enables user to add Flavour to their project with just one
parameter passed to the plugin, `flavourVersion`.